+++ example/app/dynamic-form/dynamic-form.model.ts
@@ -1,7 +1,7 @@
 import {
     DynamicFormModel,
     DynamicCheckboxModel,
-    DynamicTextInputModel,
+    DynamicInputModel,
     DynamicRadioModel,
     DynamicSelectModel,
     DynamicTextAreaModel,
@@ -12,7 +12,9 @@ export const DYNAMIC_FORM_MODEL: DynamicFormModel = new DynamicFormModel([
     new DynamicSelectModel<string>({
 
         id: "exampleOption",
-        label: "Example Option",
+        label: {
+            text: "Example Option"
+        },
         options: [
             {
                 text: "Option 1",
@@ -31,27 +33,33 @@ export const DYNAMIC_FORM_MODEL: DynamicFormModel = new DynamicFormModel([
         ]
     }),
     
-    new DynamicTextInputModel({
+    new DynamicInputModel({
 
         id: "exampleInput",
-        label: "Example Input",
+        label: {
+            text: "Example Input"
+        },
         maxLength: 51,
-        placeholder: "example input",
+        placeholder: "example input"
     }),
 
     new DynamicTextAreaModel({
 
         id: "exampleTextArea",
-        label: "Example Textarea",
+        label: {
+            hidden: true,
+            text: "Example Textarea"
+        },
         rows: 5,
         placeholder: "example Textarea",
     }),
 
     new DynamicCheckboxModel({
-
-        hideLabel: true,
+        
         id: "exampleCheckbox",
-        label: "I do agree",
+        label: {
+            text:  "I do agree",
+        },
         text: "I do agree"
     })
 ]);