+++ CHANGELOG.md
@@ -1,1 +1,8 @@
-# Changelog
+# alpha.4
+
+### **Breaking Changes**
+
+* `DynamicTextInputModel` and `DynamicNumberInputModel` condensed to `DynamicInputModel`
+* `onBlur` and `onFocus event listeners added for input and textarea controls in ui-basic and ui-bootstrap
+* `max`, `min`, `step attributes now working correctly in ui-basic and ui-bootstrap
+* `label` is now a configuration object with properties `cls`, `hidden` and `text`